refactor: Move TableFunctionExpressionWriter and AggregateExpressionRewriter to analyzer package
