+++ ksql-engine/src/main/java/io/confluent/ksql/analyzer/QueryAnalyzer.java
@@ -31,9 +31,7 @@ import io.confluent.ksql.name.FunctionName;
 import io.confluent.ksql.parser.tree.Query;
 import io.confluent.ksql.parser.tree.Sink;
 import io.confluent.ksql.serde.SerdeOption;
-import io.confluent.ksql.util.AggregateExpressionRewriter;
 import io.confluent.ksql.util.KsqlException;
-import io.confluent.ksql.util.TableFunctionExpressionRewriter;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
